Various Updates
Did some work on the loggin, Pose Estimation (mainly metrabs) and how the script checks whether Pose estimation has already been done.

Pose Estimation now runs over all available videos of a trial and not only the ones used for the given setting.

This makes sure, that all videos are run through pose estimation

+++ src/iDrink/iDrinkAnalytics.py
@@ -32,7 +32,7 @@ def use_butterworth_filter(curr_trial, data, cutoff, fs, order=4, normcutoff=Fal
     """
     from scipy.signal import butter, sosfiltfilt
 
-    order = order/2  # Order is "doubled" again by using filter 2 times
+    order = int(order/2)  # Order is "doubled" again by using filter 2 times
 
     nyquist = 0.5 * fs
